1.HTML. Create a webpage with a navbar, some main content with one main header and 4 subheaders and sections, and a footer. Also provide a list of the 10 most useful tags (e.g. div) for you in HTML
2.HTML. Research about SEO and provide a list of some of the most
important header tags that you should add to your pages for SEO

3.HTML. In the page create a grid of 3x3 like layout without using table,
or any css. Just use divs and spans

4.CSS. Create a gallery that showcases the options for the box modelLinks
to an external site. using css  For each row change the values of
border, margin and padding. For the columns try three different values
(e.g. 0px 5px 10px)

5.CSS. Create an itemized list with the same text and 6 different font
sizes, use inline styling. Choose a combination of absolute and relative
sizesLinks to an external site.. Explain on each item when would you use
that unit by means of an example

6.CSS. Add a creative part of the page that uses 4 different
positioningLinks to an external site. options. Try to think and explain
when would you use these options. Use that positioning to make sure that
your footer stays on the bottom of the screen